DelegateOpen: new library to open ~/.ssh/authorized_keys as target user
If Lukko is not allowed to access a user's ~/.ssh/authorized_keys,
spawn a new process for this user to open the file.  Even though one
could say that disallowing Lukko to access and then expecting it to
read it is unreasonable, but that's how OpenSSH does it, and what
people now expect SSH servers to do.