ssh/Connection: postpone Destroy() call until DISCONNECT is flushed
If the connection is already encrypted, we need to wait for the worker
thread to finish encrypting the DISCONNECT packet, and only after that
can we send it to the socket.

This finishes the second part of commit 62f4ca00924e8; it needs some
complicated code to ensure that the connection does no other I/O while
it's waiting for the flush.