ssh/Connection: attempt to flush the output in DoDisconnect()
When commit a597ae696b62 moved the socket send calls to class Output,
the SendPacket() method was refactored to enqueue just a buffer and
send it only when we got EPOLLOUT for the socket.  This means
SendPacket() and then Destroy() doesn't send anything.

This commit adds an Output::Flush() call after SendPacket() to
manually flush that packet.  However, this only works if the
connection is not yet encrypted; after KEX, we need to let the cipher
job finish, but that's a (more complicated) problem for another
commit.  At least KEX failures are now sent to the client.