SessionChannel: add option "exec_reject_stderr"
The correct way to reject the "exec" request for an SFTP-only account is to just return
`SSH_MSG_CHANNEL_FAILURE`.  However, this leads to a confusing error
message in the OpenSSH client.

This commit adds an option that instead replies with
`SSH_MSG_CHANNEL_SUCCESS` (even though the request has not succeeded)
and then prints an error message on the channel's STDERR.